# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: |
    Vendor Compare API for researched, metered product comparisons.

    Customer applications authenticate `/v1` operations with a tenant-scoped API key:
    `Authorization: Bearer vc_live_...`. Account, billing, and API-key management
    operations use a signed-in Clerk session and, where documented, `X-Tenant-Id`.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tenant/bootstrap:
    post:
      operationId: bootstrapTenant
      tags: [management]
      summary: Ensure a personal tenant for the signed-in Clerk user
      security: [{ ClerkBearer: [] }]
      responses:
        "200":
          description: Tenant
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tenant" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /tenant/usage:
    get:
      operationId: getTenantUsage
      tags: [management]
      summary: View tenant usage
      security: [{ ClerkBearer: [] }]
      parameters:
        - name: X-Tenant-Id
          in: header
          required: true
          description: Exact tenant being managed; the Clerk user must be an owner or admin member.
          schema: { type: string, minLength: 1, maxLength: 200 }
      responses:
        "200":
          description: Usage
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UsageSummary" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /tenant/audit:
    get:
      operationId: listTenantAudit
      tags: [management]
      summary: List tenant audit events
      security: [{ ClerkBearer: [] }]
      parameters:
        - name: X-Tenant-Id
          in: header
          required: true
          description: Exact tenant being managed; the Clerk user must be an owner or admin member.
          schema: { type: string, minLength: 1, maxLength: 200 }
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Audit events
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AuditEvent" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /tenant/api-keys:
    get:
      operationId: listTenantApiKeys
      tags: [management]
      summary: List API key metadata
      security: [{ ClerkBearer: [] }]
      parameters:
        - name: X-Tenant-Id
          in: header
          required: true
          description: Exact tenant being managed; the Clerk user must be an owner or admin member.
          schema: { type: string, minLength: 1, maxLength: 200 }
      responses:
        "200":
          description: API keys
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ApiKeyMetadata" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      operationId: createTenantApiKey
      tags: [management]
      summary: Create an API key (plaintext is returned once)
      security: [{ ClerkBearer: [] }]
      parameters:
        - name: X-Tenant-Id
          in: header
          required: true
          description: Exact tenant being managed; the Clerk user must be an owner or admin member.
          schema: { type: string, minLength: 1, maxLength: 200 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApiKeyCreateInput" }
            example:
              name: Production integration
              scopes: [comparisons:read, comparisons:write, usage:read]
      responses:
        "201":
          description: Created API key
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiKeyCreated" }
              example:
                id: 7
                name: Production integration
                prefix: vc_live_ab12
                scopes: [comparisons:read, comparisons:write, usage:read]
                expiresAt: null
                revokedAt: null
                lastUsedAt: null
                createdAt: "2026-09-15T01:00:00.000Z"
                key: vc_live_ab12_example_returned_once
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /tenant/api-keys/{id}/rotate:
    post:
      operationId: rotateTenantApiKey
      tags: [management]
      summary: Revoke an API key and issue a replacement
      security: [{ ClerkBearer: [] }]
      parameters:
        - name: X-Tenant-Id
          in: header
          required: true
          description: Exact tenant being managed; the Clerk user must be an owner or admin member.
          schema: { type: string, minLength: 1, maxLength: 200 }
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "201":
          description: Replacement key, plaintext returned once
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiKeyCreated" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /tenant/api-keys/{id}/revoke:
    post:
      operationId: revokeTenantApiKey
      tags: [management]
      summary: Revoke an API key
      security: [{ ClerkBearer: [] }]
      parameters:
        - name: X-Tenant-Id
          in: header
          required: true
          description: Exact tenant being managed; the Clerk user must be an owner or admin member.
          schema: { type: string, minLength: 1, maxLength: 200 }
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204": { description: Revoked }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /stripe/checkout:
    post:
      operationId: createStripeCheckout
      tags: [management]
      summary: Create a hosted Stripe checkout
      security: [{ ClerkBearer: [] }]
      parameters:
        - name: X-Tenant-Id
          in: header
          required: true
          description: Exact tenant being managed; the Clerk user must be an owner or admin member.
          schema: { type: string, minLength: 1, maxLength: 200 }
      description: |
        Creates a Stripe-hosted subscription payment page for the signed-in tenant administrator.
        Redirect the customer to `purchaseUrl`. The redirect does not activate access:
        the server verifies a successful non-refunded payment, active subscription, and
        current billing period before enabling API keys. Returns `503 billing_not_configured`
        until a verified Stripe recurring price is configured.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CheckoutInput" }
            example:
              redirectUrl: https://your-app.example.com/billing/complete
      responses:
        "201":
          description: Hosted checkout
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CheckoutResponse" }
              example:
                purchaseUrl: https://checkout.stripe.com/c/pay/example
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "503": { $ref: "#/components/responses/ConfigurationError" }
  /tenant/stripe/reconcile:
    post:
      operationId: reconcileStripeBilling
      tags: [management]
      summary: Reconcile a stored Stripe checkout server-side
      security: [{ ClerkBearer: [] }]
      parameters:
        - name: X-Tenant-Id
          in: header
          required: true
          description: Exact tenant being managed; the Clerk user must be an owner or admin member.
          schema: { type: string, minLength: 1, maxLength: 200 }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BillingReconciliationInput" }
      responses:
        "200":
          description: Verified or unverified billing status
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BillingReconciliation" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "503": { $ref: "#/components/responses/ConfigurationError" }
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: Vendor Compare API key
      description: "Tenant-scoped API key. Send Authorization: Bearer vc_live_.... Scope is documented per operation. The plaintext key is returned only when created or rotated."
    ClerkBearer:
      type: http
      scheme: bearer
      bearerFormat: Clerk session
      description: Signed-in Clerk session token for management operations only; do not use this credential for /v1 calls.
  parameters:
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
...
  responses:
    BadRequest:
      description: Invalid request
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
    Unauthorized:
      description: Missing or invalid authentication
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
    Forbidden:
      description: Authentication lacks the required scope or tenant role
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
    NotFound:
      description: Resource not found
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
...
    ConfigurationError:
      description: Required server configuration is not available
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
...
  schemas:
...
    ErrorResponse:
      type: object
      description: Stable structured error envelope; error is retained for legacy clients.
      properties:
        code: { type: string, example: invalid_request }
        message: { type: string }
        requestId: { type: string }
        details: { type: object, additionalProperties: true }
      required: [code, message]
    Tenant:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        plan: { type: string }
        billingStatus: { type: string, enum: [inactive, active, past_due] }
        includedComparisons: { type: integer }
        requestsPerMinute: { type: integer }
      required: [id, name, plan, billingStatus, includedComparisons, requestsPerMinute]
    BillingReconciliation:
      type: object
      properties:
        status: { type: string, enum: [active, unverified] }
        subscriptionId: { type: string, nullable: true }
      required: [status]
    BillingReconciliationInput:
      type: object
      properties:
        checkoutSessionId: { type: string, minLength: 1, maxLength: 255 }
    AuditEvent:
      type: object
      properties:
        id: { type: integer }
        actorId: { type: string }
        action: { type: string }
        metadata: { type: object, additionalProperties: true }
        createdAt: { type: string, format: date-time }
      required: [id, actorId, action, metadata, createdAt]
    ApiKeyCreateInput:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        scopes:
          type: array
          items: { type: string, enum: [comparisons:read, comparisons:write, usage:read] }
        expiresAt: { type: string, format: date-time }
      required: [name]
    ApiKeyMetadata:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        prefix: { type: string }
        scopes: { type: array, items: { type: string } }
        expiresAt: { type: string, format: date-time, nullable: true }
        revokedAt: { type: string, format: date-time, nullable: true }
        lastUsedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
      required: [id, name, prefix, scopes, createdAt]
    ApiKeyCreated:
      allOf:
        - $ref: "#/components/schemas/ApiKeyMetadata"
        - type: object
          properties:
            key: { type: string, description: Plaintext key; returned only once. }
          required: [key]
    UsageSummary:
      type: object
      properties:
        month: { type: string, example: "2025-01-15/2025-02-15", description: Provider period label; inactive/bootstrap tenants use YYYY-MM. }
        periodStart: { type: string, format: date-time }
        periodEnd: { type: string, format: date-time }
        included: { type: integer }
        used: { type: integer }
        remaining: { type: integer, description: Comparisons remaining before the prepaid hard cap is reached. }
        exhausted: { type: boolean }
      required: [month, periodStart, periodEnd, included, used, remaining, exhausted]
    CheckoutInput:
      type: object
      properties:
        redirectUrl: { type: string, format: uri }
      required: [redirectUrl]
    CheckoutResponse:
      type: object
      properties:
        purchaseUrl: { type: string, format: uri }
      required: [purchaseUrl]
...